Move: 영상 게시판 css 파일 하나로 이동
